refactor: 목업 시스템을 window에서 localStorage로 교체
